started with planning model.
start on main planning workflow. Since this is an MVP, we will still keep all the code in one file - but the code should be logical enough to seprate into different modules or files when get to move from MVP to launch stage. I am going to describe at a very high level what planning entails. I would like you to synthesize these requirements and then propose an implementation plan we can iterate on and build the incremental planning capabilities.



The goal of planning is to fit as many initiatives into the delivery plan for the year, as we have available capacity. Capacity is constrained by BIS (builders in seats) [realistic scenario]. The ideal however is for the teams to reach fully funded state as soon as possible (ideally we need to model a capacity resource forecasting model, this will come at a later increment).



We first need to create the high level year plan of initiatives, call them intiatives or workstreams. Each initiative impacts one or more service teams (2P teams). We need a way of capturing the initiative line items and assign the impacted teams. Since this is a high level year plan, we don't need to show which specific service is impacted, but we need to capture the development estimate for that team, by using headcount for the year, using "SDE Years". Example: If an estimate value is "1", this implies it would take one engineer capacity for the full year to work on that initiative. Since we know the full resource capacity (number of engineers in the organisation), we can quickly see how capacity is being used up. The list of initiatives might be large such that the organization runs out of capacity (either BIS which is a yellow flag), or runs out of full funded capacity (red flag). All initiatives that can be contained within the available funded capacity are considered "Above the line" ATL, which means the organization can commit to delivery. All other initiatves fall "Below the line" BTL because capacity runs out. This initial planning view must be implemented as an interactive list, where rows can be moved around dynamically to get initiatives ATL or BTL, as a way to prioritize the urgent / important initiatives. Provision must be made to protect some initiatives - i.e. they are so important that cannot be moved down in the list, the only option is to reduce headcount (capacity) estimates, or reduce scope. What I'm describing is standard software engineering management practice, you can research the industry practices - and suggest other things to consider.